Simple Missing Number Finder
This is the first part in making the program go through all possible combinations while solving the sudoku board. In this part I made a simple program that goes through the first column of the board and findes the missing numbers. Also I started the part of the program that makes all of the missing numbre combinations.